requirements.txt
omegaconf
loguru
smplx
opencv-python
tqdm
pyyaml
configargparse
pose-format
tensorflow
ffmpeg-python
bpy
